refactor(core/rdma): tidy header names/placement
Two cosmetic cleanups, no behavior change:
- providers/utils.h holds only provider-agnostic helpers (WcStatusString,
  get_num_wqes_in_atomic), so move it up out of providers/ and give it the C++
  extension: providers/utils.h -> transport/rdma/utils.hpp.
- primitives.hpp defines just the host IBVerbsHandle wrapper -> rename to
  ibverbs_handle.hpp so the name matches its content.

Includers updated. Full build (BNXT) clean.
